Move facet search field components
